Moved all of the custom structs into their own package
