+++ prototypes/rust/postboard/backend/src/main.rs
@@ -2,8 +2,6 @@
 extern crate rocket;
 
 mod routes;
-mod serde;
-mod structs;
 
 /// Launches the Rocket application
 #[launch]
