+++ prototypes/rust/postboard/types/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "postboard-backend"
+name = "postboard-types"
 version = "0.1.0"
 edition = "2021"
 
@@ -7,6 +7,4 @@ edition = "2021"
 
 [dependencies]
 chrono = "0.4.31"
-postboard-types = {path = "../types"}
-rocket = { version = "0.5.0", features = ["serde_json", "json"] }
-serde = "1.0.195"
+serde = { version = "1.0.195", features = ["derive"] }
