+++ prototypes/rust/postboard/types/src/lib.rs
@@ -1,0 +1,7 @@
+mod serde;
+
+mod post;
+pub use post::Post;
+
+mod submission;
+pub use submission::Submission;
